add _form partial for create and edit template

+++ src/Controller/PinsController.php
@@ -49,7 +49,7 @@ class PinsController extends AbstractController
         }
 
         return $this->render('pins/create.html.twig', [
-            'formCreate' => $form->createView()
+            'form' => $form->createView()
 
         ]);
 
@@ -98,10 +98,15 @@ return $this->render('pins/show.html.twig', compact('pin'));
      * @route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="DELETE")
      */
 
-public function delete(Pin $pin, EntityManagerInterface $em): Response
+public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
 {
-$em->remove($pin);
-$em->flush();
+
+    if ($this->isCsrfTokenValid('pin_deletion_'. $pin->getId(), $request->request->get('csrf_token'))) {
+        $em->remove($pin);
+        $em->flush();
+    }
+
+
 
             return $this->redirectToRoute('app_home');
